feat: regenerate nginx config (in theory)

+++ orchestrator/nginx.py
@@ -72,12 +72,14 @@ def build_upstreams(state: OrchestratorState) -> dict[str, str]:
 
 
 def regenerate_nginx_config(state: OrchestratorState, reload: bool):
+    print(f"Regenerating nginx config...")
     conf_folder = Path("nginx/conf.d")
     if not conf_folder.exists():
         conf_folder.mkdir()
-    for file in conf_folder.iterdir():
-        if file.is_file():
-            file.unlink()
+    if not reload:
+        for file in conf_folder.iterdir():
+            if file.is_file():
+                file.unlink()
 
     with open(conf_folder / "base.conf", "w") as file:
         file.write(build_nginx_config(state))
